# [Merged by Bors] - Optimise tree hash caching for block production