@michaelsproul 2020-12-20
This updated graph from `@potuz` shows this PR (ac59dfa) clearly outperforming v1.0.4 (1abc70e). The other commits were intermediate versions that I ran while testing.

![lighthouse1](https://user-images.githubusercontent.com/4452260/102727478-f8404f00-4379-11eb-995f-83d6ba00650a.jpeg)